refactor: common 모듈 해체
- common.shffule -> shffule
- common.PositiveNumber -> math.PositiveNumber
