+++ src/main/kotlin/lotto/domain/analysis/MatchCondition.kt
@@ -1,6 +1,6 @@
 package lotto.domain.analysis
 
-import common.PositiveNumber
+import math.PositiveNumber
 
 data class MatchCondition(
     val matchSuccessCount: PositiveNumber,
