+++ src/main/kotlin/lotto/view/LottoInputView.kt
@@ -1,6 +1,6 @@
 package lotto.view
 
-import common.PositiveNumber
+import math.PositiveNumber
 import lotto.domain.lottonumber.LottoNumbers
 
 interface LottoInputView {
