+++ src/main/kotlin/lotto/view/LottoStarter.kt
@@ -1,6 +1,6 @@
 package lotto.view
 
-import common.shffule.RandomShuffler
+import shffule.RandomShuffler
 import lotto.domain.analysis.RealLottoResultAnalyst
 import lotto.domain.shop.RealLottoShop
 import lotto.domain.shop.machine.RealLottoGameMachine
