+++ src/main/kotlin/shffule/Shuffler.kt
@@ -1,4 +1,4 @@
-package common.shffule
+package shffule
 
 interface Shuffler<T> {
