+++ src/test/kotlin/calculator/StringParserTest.kt
@@ -1,6 +1,6 @@
 package calculator
 
-import common.PositiveNumber
+import math.PositiveNumber
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.DisplayName
