+++ src/test/kotlin/lotto/domain/shop/RealLottoShopTest.kt
@@ -1,6 +1,6 @@
 package lotto.domain.shop
 
-import common.PositiveNumber
+import math.PositiveNumber
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import lotto.domain.lottonumber.LottoNumber
