+++ src/test/kotlin/math/PositiveNumberTest.kt
@@ -1,4 +1,4 @@
-package common
+package math
 
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
